Removed Example Sessions, Added change, edit, delete handlers. Added API call to populate sessions.
Big update. Removed old test sessions and added call to API to fetch sessions from SQl server and populate the list. This is called on mount and when deleting a session. Also added in Change handler for edit field use. Implemented delte and edit handlers to call API to edit the database and update the state of the app. Current issue with the editing not being super friendly, but we will get there :).